# Use PathElementSet and similar to index PathElement