Currently, we index PathElement by turning them into strings, which is neither fast nor accurate.
We can simply use the `PathElementSet` when we need a set, and we can build a map of PathElement to Value when needed using very similar code.

Improvement:
```
benchmark                                              old ns/op     new ns/op     delta
BenchmarkOperations/Pod/Update-12                      187994        162602        -13.51%
BenchmarkOperations/Pod/Apply-12                       525375        452963        -13.78%
BenchmarkOperations/Node/Update-12                     273261        220959        -19.14%
BenchmarkOperations/Node/Apply-12                      760287        667991        -12.14%
BenchmarkOperations/Endpoints/Update-12                32649         32794         +0.44%
BenchmarkOperations/Endpoints/Apply-12                 2072047       1862396       -10.12%
BenchmarkFromUnstructured/Pod-12                       202893        182379        -10.11%
BenchmarkFromUnstructured/Node-12                      331193        308072        -6.98%
BenchmarkFromUnstructured/Endpoints-12                 8057365       7754632       -3.76%

benchmark                                              old allocs     new allocs     delta
BenchmarkOperations/Pod/Update-12                      445            284            -36.18%
BenchmarkOperations/Pod/Apply-12                       1179           771            -34.61%
BenchmarkOperations/Node/Update-12                     601            336            -44.09%
BenchmarkOperations/Node/Apply-12                      1876           1138           -39.34%
BenchmarkOperations/Endpoints/Update-12                74             74             +0.00%
BenchmarkOperations/Endpoints/Apply-12                 3142           1120           -64.35%
BenchmarkFromUnstructured/Pod-12                       477            392            -17.82%
BenchmarkFromUnstructured/Node-12                      1107           900            -18.70%
BenchmarkFromUnstructured/Endpoints-12                 21115          19089          -9.60%
```