@apelisse 2019-10-03
More or less replaces #112