@apelisse 2019-10-07
I think we should probably re-use the code in other places rather than add another implementation of the set, but I don't have a lot of time for that right now so I've decided to add a test instead.
I can create an issue to track that if needed. And I'll add a comment.