@apelisse 2019-10-08
OK, I've added the pre-allocation of the right size for the arrays, here are the new numbers, pretty good:
```
benchmark                                              old ns/op     new ns/op     delta
BenchmarkDeducedSimple-12                              113151        114199        +0.93%
BenchmarkDeducedNested-12                              319775        296626        -7.24%
BenchmarkDeducedNestedAcrossVersion-12                 329968        328015        -0.59%
BenchmarkLeafConflictAcrossVersion-12                  114734        107390        -6.40%
BenchmarkMultipleApplierRecursiveRealConversion-12     1565560       1588022       +1.43%
BenchmarkOperations/Pod/Update-12                      187994        115494        -38.57%
BenchmarkOperations/Pod/Apply-12                       525375        294227        -44.00%
BenchmarkOperations/Node/Update-12                     273261        172191        -36.99%
BenchmarkOperations/Node/Apply-12                      760287        502819        -33.86%
BenchmarkOperations/Endpoints/Update-12                32649         22770         -30.26%
BenchmarkOperations/Endpoints/Apply-12                 2072047       1003676       -51.56%
BenchmarkFromUnstructured/Pod-12                       202893        115718        -42.97%
BenchmarkFromUnstructured/Node-12                      331193        234963        -29.06%
BenchmarkFromUnstructured/Endpoints-12                 8057365       6069791       -24.67%

benchmark                                              old allocs     new allocs     delta
BenchmarkDeducedSimple-12                              552            552            +0.00%
BenchmarkDeducedNested-12                              1534           1498           -2.35%
BenchmarkDeducedNestedAcrossVersion-12                 1602           1566           -2.25%
BenchmarkLeafConflictAcrossVersion-12                  532            532            +0.00%
BenchmarkMultipleApplierRecursiveRealConversion-12     5229           5232           +0.06%
BenchmarkOperations/Pod/Update-12                      445            282            -36.63%
BenchmarkOperations/Pod/Apply-12                       1179           744            -36.90%
BenchmarkOperations/Node/Update-12                     601            342            -43.09%
BenchmarkOperations/Node/Apply-12                      1876           1103           -41.20%
BenchmarkOperations/Endpoints/Update-12                74             72             -2.70%
BenchmarkOperations/Endpoints/Apply-12                 3142           1103           -64.89%
BenchmarkFromUnstructured/Pod-12                       477            370            -22.43%
BenchmarkFromUnstructured/Node-12                      1107           854            -22.85%
BenchmarkFromUnstructured/Endpoints-12                 21115          19076          -9.66%
```