@lavalamp 2019-10-22
/lgtm
/approve

Sorry, didn't realize this was waiting on me.